Move mania-specific enums inside `LegacyManiaSkinConfiguration`
