--- osu.Game/Skinning/LegacyComboBurstStyle.cs
@@ -1,12 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-namespace osu.Game.Skinning
-{
-    public enum LegacyComboBurstStyle
-    {
-        Left = 0,
-        Right = 1,
-        Both = 2,
-    }
-}
