+++ osu.Game/Skinning/LegacySkin.cs
@@ -197,12 +197,12 @@ namespace osu.Game.Skinning
                 case LegacyManiaSkinConfigurationLookups.NoteBodyStyle:
 
                     if (existing.NoteBodyStyle != null)
-                        return SkinUtils.As<TValue>(new Bindable<LegacyNoteBodyStyle>(existing.NoteBodyStyle.Value));
+                        return SkinUtils.As<TValue>(new Bindable<LegacyManiaSkinConfiguration.LegacyNoteBodyStyle>(existing.NoteBodyStyle.Value));
 
                     if (GetConfig<SkinConfiguration.LegacySetting, decimal>(SkinConfiguration.LegacySetting.Version)?.Value < 2.5m)
-                        return SkinUtils.As<TValue>(new Bindable<LegacyNoteBodyStyle>(LegacyNoteBodyStyle.Stretch));
+                        return SkinUtils.As<TValue>(new Bindable<LegacyManiaSkinConfiguration.LegacyNoteBodyStyle>());
 
-                    return SkinUtils.As<TValue>(new Bindable<LegacyNoteBodyStyle>(LegacyNoteBodyStyle.RepeatBottom));
+                    return SkinUtils.As<TValue>(new Bindable<LegacyManiaSkinConfiguration.LegacyNoteBodyStyle>(LegacyManiaSkinConfiguration.LegacyNoteBodyStyle.RepeatBottom));
 
                 case LegacyManiaSkinConfigurationLookups.NoteImage:
                     Debug.Assert(maniaLookup.ColumnIndex != null);
