--- osu.Game/Skinning/LegacySpecialStyle.cs
@@ -1,12 +1,0 @@
-// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
-// See the LICENCE file in the repository root for full licence text.
-
-namespace osu.Game.Skinning
-{
-    public enum LegacySpecialStyle
-    {
-        None = 0,
-        Left = 1,
-        Right = 2,
-    }
-}
